Refactor the comment section
