+++ web/src/components/WorkRequestCommentSection/WorkRequestCommentSection.test.tsx
@@ -1,0 +1,14 @@
+import { render } from '@redwoodjs/testing/web'
+
+import WorkRequestCommentSection from './WorkRequestCommentSection'
+
+//   Improve this test with help from the Redwood Testing Doc:
+//    https://redwoodjs.com/docs/testing#testing-components
+
+describe('WorkRequestCommentSection', () => {
+  it('renders successfully', () => {
+    expect(() => {
+      render(<WorkRequestCommentSection />)
+    }).not.toThrow()
+  })
+})
